+++ backend/src/main/java/com/be/web/dto/response/WorkshopListDTO.java
@@ -1,5 +1,6 @@
 package com.be.web.dto.response;
 
+import com.be.web.dto.TeacherInfoDTO;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -16,7 +17,7 @@ public class WorkshopListDTO {
     private Long id;
     private String title;
     private String shortDescription;
-    private UserBasicDTO teacher;
+    private TeacherInfoDTO teacher;
     private LocalDate startDate;
     private LocalDate endDate;
     private BigDecimal price;